# packages/userManagement/swagger.yaml
openapi: 3.0.3
info:
  title: iot-callbot-node
  description: |-

  contact:
    email: nguyenduyphuong3103@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:8000/api/
...
paths:
  /user/register:
    post:
      tags:
        - user
      summary: Register
      description: Users do not have an account yet
      operationId: registerUser
      requestBody:
        description: Register user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        default:
          description: successful registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "Register successfully!!!"
                  resData:
                    $ref: "#/components/schemas/SentMail"
  /user/confirmRegisterUser/{confirmUserToken}:
    get:
      tags:
        - user
      summary: confirm register
      description: "Confirm user registration via email"
      operationId: confirmRegisterUser
      parameters:
        - name: confirmUserToken
          in: path
          description: "The confirmUserToken that needs to be fetched. "
          required: true
          schema:
            type: string
      responses:
        default:
          description: Account confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "Account confirmed successfully !!!"
                  resData:
                    $ref: "#/components/schemas/ConfirmUser"
  /user/login:
    post:
      tags:
        - user
      summary: Log in
      description: "Logs user into the system"
      operationId: loginUser
      requestBody:
        description: Login user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Login"
        required: true
      responses:
        default:
          description: Account confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "Logged in successfully!!!"
                  resData:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImUzOTM2MTY0LTdjYWYtNDgwZC05NDlmLTdkODIyZWRlNGE2MSIsInJvbGUiOiJ1c2VyIiwiaWF0IjoxNzE2MTc2OTQ3LCJleHAiOjE3MTYxODA1NDd9.Mn-iR79rUpX4R-lj1y5JpqQCLHQiFNrg-CwuJHMCC4o"
                      isEmailExist:
                        $ref: "#/components/schemas/ConfirmUser"
  /user/logout:
    get:
      tags:
        - user
      summary: Log out
      description: "Logs out current logged in user session"
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        default:
          description: Account confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "You have successfully logged out !!!"
...
  /user/changePassword:
    patch:
      tags:
        - user
      summary: Change password
      description: Change password user to server
      operationId: Change user password
      security:
        - BearerAuth: []
      requestBody:
        description: Change password user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePassword"
        required: true
      responses:
        default:
          description: Changed refresh token
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "Changed password successfully!!!"
  /user/forgotPassword:
    get:
      tags:
        - user
      summary: Forgot password
      description: "Forgot password user"
      operationId: forgotPassword
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
          description: Email of user need to reset password
      responses:
        default:
          description: successful registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  resData:
                    $ref: "#/components/schemas/SentMail"
  /user/verifyForgotPassword/{resetToken}:
    get:
      tags:
        - user
      summary: Verify forgot password
      description: "Check the token to authenticate the user"
      operationId: verifyForgotPassword
      parameters:
        - name: resetToken
          in: path
          description: "The resetToken that needs to be fetched. "
          required: true
          schema:
            type: string
      responses:
        default:
          description: Valid OTP
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: Valid OTP
                  resData:
                    type: string
                    example: "e3936164-7caf-480d-949f-7d822ede4a61"
  /user/resetPassword/{id}:
    put:
      tags:
        - user
      summary: Reset password
      description: Reset password user
      operationId: resetPassword
      parameters:
        - name: id
          in: path
          description: "id of user"
          required: true
          schema:
            type: string
      requestBody:
        description: new password
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  example: "777"
        required: true
      responses:
        default:
          description: Account confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "Password changed successfully !!!"
                  resData:
                    $ref: "#/components/schemas/UserRegister"
  /user/updateProfile:
    put:
      tags:
        - user
      summary: Reset password
      description: Reset password user
      operationId: resetPassword
      parameters:
        - name: id
          in: path
          description: "id of user"
          required: true
          schema:
            type: string
      requestBody:
        description: new password
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  example: "777"
        required: true
      responses:
        default:
          description: Account confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "Password changed successfully !!!"
                  resData:
                    $ref: "#/components/schemas/UserRegister"
...
components:
  schemas:
...
    UserRegister:
      type: object
      properties:
        email:
          type: string
          example: "nguyenduyphuong3103@gmail.com"
        name:
          type: string
          example: "Nguyen Duy Phuong"
        password:
          type: string
          example: "123456"
        passwordConfirm:
          type: string
          example: "123456"
      xml:
        name: user
...
    ChangePassword:
      type: object
      properties:
        currentPassword:
          type: string
          example: 000
        newPassword:
          type: string
          example: 777
    Login:
      type: object
      properties:
        email:
          type: string
          example: "nguyenduyphuong310320@gmail.com"
        password:
          type: string
          example: "000"
      xml:
        name: user
    ConfirmUser:
      type: object
      properties:
        id:
          type: string
          example: "e3936164-7caf-480d-949f-7d822ede4a61"
        name:
          type: string
          example: "nguyenduyphuong3103"
        avatar:
          type: string
          example: null
        email:
          type: string
          example: "nguyenduyphuong310320@gmail.com"
        isLocked:
          type: boolean
          example: false
        createdAt:
          type: string
          example: "2024-05-20T03:31:40.769Z"
        updatedAt:
          type: string
          example: "2024-05-20T03:32:35.442Z"
      xml:
        name: user
...
    SentMail:
      type: object
      properties:
        accepted:
          type: array
          items:
            type: string
          example:
            - nguyenduyphuong3103200@gmail.com
        rejected:
          type: array
          items:
            type: string
          example:
            -
        ehlo:
          type: array
          items:
            type: string
          example:
            - SIZE 35882577
            - 8BITMIME
            - AUTH LOGIN PLAIN XOAUTH2 PLAIN-CLIENTTOKEN OAUTHBEARER XOAUTH
            - ENHANCEDSTATUSCODES
            - PIPELINING
            - CHUNKING
            - SMTPUTF8
        envelopeTime:
          type: number
          example: 858
        messageTime:
          type: number
          example: 830
        messageSize:
          type: number
          example: 758
        response:
          type: string
          example:
            - 250 2.0.0 OK  1715762165 98e67ed59e1d1-2b62863a5b2sm13149255a91.6 - gsmtp
        envelope:
          type: object
          properties:
            from:
              type: string
              example: no-reply@iot-callbot-node.com
            to:
              type: array
              items:
                type: string
              example:
                - nguyenduyphuong3103200@gmail.com
        messageId:
          type: string
          example: <93c2adde-ce12-ee51-6fe5-41d8934660b1@iot-callbot-node.com>
      xml:
        name: user
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT